Add a load method
I factored out some common path manipulation so it could be shared between
the load and save methods.

I implemented this in the simplest possible way using plain strings.
This will cover my primary use case, but a future version of the module will
be more generically useful if is supports reading and writing streams
and handles character encoding properly.